Fix a mis-implementation in ALPSO.
The global Langrangian was updated during updating each particles
in a swarm, which means they could use different gbest_L if gbest_L
was updated in the early phase of the for loop. Now, the updation
of gbest_L is in a different loop.

Increase the default dtol from 0.05 to 0.1