set workers in Beamline and more exceptions.
In order to run the parallel version of a simulation code, one
should set 'workers' for the corresponding Beamline. Previously
it was set in Operation (LinacOptimization or LinacJitter). A
workers property is also added for Optimizer, which paves the way
for adding parallelized Optimizer in the future.

Add more Exceptions to provide more informative error message.

Add more checks before running external code.